PKE on AWS: add cloudformation stack failure events to the error message returned in case of failure
